+++ examples/platformio/src/main.cpp
@@ -323,7 +323,9 @@ void setup()
       PsychicHttpServer* redirectServer = new PsychicHttpServer();
       redirectServer->config.ctrl_port = 20424; // just a random port different from the default one
       redirectServer->onNotFound([](PsychicRequest* request, PsychicResponse* response) {
-        String url = "https://" + request->host() + request->url();
+        String url = "https://";
+        url += request->host();
+        url += request->url();
         return response->redirect(url.c_str()); });
     }
 #endif
@@ -413,7 +415,8 @@ void setup()
     // ip - get info about the client
     //  curl -i http://psychic.local/ip
     server.on("/ip", HTTP_GET, [](PsychicRequest* request, PsychicResponse* response) {
-      String output = "Your IP is: " + request->client()->remoteIP().toString();
+      String output = "Your IP is: ";
+      output += request->client()->remoteIP().toString();
       return response->send(output.c_str());
     });
 
@@ -441,7 +444,7 @@ void setup()
 
       // work with some params
       if (request->hasParam("foo")) {
-        String foo = request->getParam("foo")->value();
+        String foo = request->getParam("foo", "");
         output["foo"] = foo;
       }
 
@@ -477,9 +480,15 @@ void setup()
             std::smatch matches;
             if (request->getRegexMatches(matches)) {
               String output;
-              output += "Matches: " + String(matches.size()) + "<br/>\n";
-              output += "Matched URI: " + String(matches.str(0).c_str()) + "<br/>\n";
-              output += "Match 1: " + String(matches.str(1).c_str()) + "<br/>\n";
+              output += "Matches: ";
+              output += matches.size();
+              output += "<br/>\n";
+              output += "Matched URI: ";
+              output += matches.str(0).c_str();
+              output += "<br/>\n";
+              output += "Match 1: ";
+              output += matches.str(1).c_str();
+              output += "<br/>\n";
 
               return response->send(output.c_str());
             } else
@@ -546,8 +555,12 @@ void setup()
     //  curl -F "param1=value1" -F "param2=value2" -X POST http://psychic.local/post
     server.on("/post", HTTP_POST, [](PsychicRequest* request, PsychicResponse* response) {
       String output;
-      output += "Param 1: " + request->getParam("param1")->value() + "<br/>\n";
-      output += "Param 2: " + request->getParam("param2")->value() + "<br/>\n";
+      output += "Param 1: ";
+      output += request->getParam("param1", "");
+      output += "<br/>\n";
+      output += "Param 2: ";
+      output += request->getParam("param2", "");
+      output += "<br/>\n";
 
       return response->send(output.c_str());
     });
@@ -588,8 +601,13 @@ void setup()
 
     // gets called after upload has been handled
     uploadHandler->onRequest([](PsychicRequest* request, PsychicResponse* response) {
-      String url = "/" + request->getFilename();
-      String output = "<a href=\"" + url + "\">" + url + "</a>";
+      String url = "/";
+      url += request->getFilename();
+      String output = "<a href=\"";
+      output += url;
+      output += "\">";
+      output += url;
+      output += "</a>";
 
       return response->send(output.c_str());
     });
@@ -634,15 +652,28 @@ void setup()
       if (request->hasParam("file_upload")) {
         PsychicWebParameter* file = request->getParam("file_upload");
 
-        String url = "/" + file->value();
-        output += "<a href=\"" + url + "\">" + url + "</a><br/>\n";
-        output += "Bytes: " + String(file->size()) + "<br/>\n";
+        String url = "/";
+        url += file->value();
+        output += "<a href=\"";
+        output += url;
+        output += "\">";
+        output += url;
+        output += "</a><br/>\n";
+        output += "Bytes: ";
+        output += file->size();
+        output += "<br/>\n";
       }
 
-      if (request->hasParam("param1"))
-        output += "Param 1: " + request->getParam("param1")->value() + "<br/>\n";
-      if (request->hasParam("param2"))
-        output += "Param 2: " + request->getParam("param2")->value() + "<br/>\n";
+      if (request->hasParam("param1")) {
+        output += "Param 1: ";
+        output += request->getParam("param1", "");
+        output += "<br/>\n";
+      }
+      if (request->hasParam("param2")) {
+        output += "Param 2: ";
+        output += request->getParam("param2", "");
+        output += "<br/>\n";
+      }
 
       return response->send(output.c_str());
     });
@@ -657,10 +688,16 @@ void setup()
     PsychicUploadHandler* multipartFormHandler = new PsychicUploadHandler();
     multipartFormHandler->onRequest([](PsychicRequest* request, PsychicResponse* response) {
       String output;
-      if (request->hasParam("param1"))
-        output += "Param 1: " + request->getParam("param1")->value() + "<br/>\n";
-      if (request->hasParam("param2"))
-        output += "Param 2: " + request->getParam("param2")->value() + "<br/>\n";
+      if (request->hasParam("param1")) {
+        output += "Param 1: ";
+        output += request->getParam("param1", "");
+        output += "<br/>\n";
+      }
+      if (request->hasParam("param2")) {
+        output += "Param 2: ";
+        output += request->getParam("param2", "");
+        output += "<br/>\n";
+      }
 
       return response->send(output.c_str());
     });
@@ -695,7 +732,9 @@ void setup()
     // 404:   curl -F "foo=bar" http://psychic.local/post-filter
     server.on("/post-filter", HTTP_POST, [](PsychicRequest* request, PsychicResponse* response) {
             String output;
-            output += "Secret: " + request->getParam("secret")->value() + "<br/>\n";
+            output += "Secret: ";
+            output += request->getParam("secret", "");
+            output += "<br/>\n";
 
             return response->send(output.c_str());
           })