+++ src/PsychicEndpoint.cpp
@@ -35,9 +35,9 @@ PsychicHandler* PsychicEndpoint::handler()
   return _handler;
 }
 
-String PsychicEndpoint::uri()
+const char* PsychicEndpoint::uri()
 {
-  return _uri;
+  return _uri.c_str();
 }
 
 esp_err_t PsychicEndpoint::requestCallback(httpd_req_t* req)
@@ -135,6 +135,6 @@ esp_err_t PsychicEndpoint::process(PsychicRequest* request)
   esp_err_t ret = ESP_ERR_HTTPD_INVALID_REQ;
   if (_handler != NULL)
     ret = _handler->process(request);
-  ESP_LOGD(PH_TAG, "Endpoint %s processed %s: %s", _uri.c_str(), request->uri().c_str(), esp_err_to_name(ret));
+  ESP_LOGD(PH_TAG, "Endpoint %s processed %s: %s", _uri.c_str(), request->uri(), esp_err_to_name(ret));
   return ret;
 }