+++ src/PsychicResponse.h
@@ -14,7 +14,7 @@ class PsychicResponse
     int _code;
     char _status[60];
     std::list<HTTPHeader> _headers;
-    String _contentType;
+    std::string _contentType;
     int64_t _contentLength;
     const char* _body;
 
@@ -26,9 +26,9 @@ class PsychicResponse
 
     void setCode(int code);
     int getCode() { return _code; }
-    
+
     void setContentType(const char* contentType);
-    String& getContentType() { return _contentType; }
+    const char* getContentType() { return _contentType.c_str(); }
 
     void setContentLength(int64_t contentLength) { _contentLength = contentLength; }
     int64_t getContentLength(int64_t contentLength) { return _contentLength; }
@@ -74,7 +74,7 @@ class PsychicResponseDelegate
     void setCode(int code) { _response->setCode(code); }
 
     void setContentType(const char* contentType) { _response->setContentType(contentType); }
-    String& getContentType() { return _response->getContentType(); }
+    const char* getContentType() { return _response->getContentType(); }
 
     void setContentLength(int64_t contentLength) { _response->setContentLength(contentLength); }
     int64_t getContentLength(int64_t contentLength) { return _response->getContentLength(); }